Comentados los endpoints no utilizados por el momento
Creación de carpeta "pojo" para contener la clase PuestoFiltro
