+++ src/main/java/com/evalia/backEntrevistasInformes/controller/EntrevistasController.java
@@ -10,7 +10,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.evalia.backEntrevistasInformes.model.entity.EntrevistaEntity;
 import com.evalia.backEntrevistasInformes.model.entrevista.EntrevistaFinalizadaDTO;
 import com.evalia.backEntrevistasInformes.service.entrevistas.IEntrevistasService;
 import com.evalia.backEntrevistasInformes.service.survey.ICategoriaTecnologiaService;
