+++ src/main/java/com/evalia/backEntrevistasInformes/model/pojo/PuestoFiltro.java
@@ -1,0 +1,22 @@
+package com.evalia.backEntrevistasInformes.model.pojo;
+
+public class PuestoFiltro {
+    private String tecnologia;
+    private String nivel;
+
+    public String getTecnologia() {
+        return tecnologia;
+    }
+
+    public void setTecnologia(String tecnologia) {
+        this.tecnologia = tecnologia;
+    }
+
+    public String getNivel() {
+        return nivel;
+    }
+
+    public void setNivel(String nivel) {
+        this.nivel = nivel;
+    }
+}
